+++ src/modules/script.cpp
@@ -9,9 +9,11 @@ namespace modules {
   script_module::script_module(const bar_settings& bar, string name_)
       : module<script_module>(bar, move(name_))
       , m_tail(m_conf.get(name(), "tail", false))
-      , m_interval(m_conf.get<script_runner::interval>(name(), "interval", m_tail ? 0s : 5s))
+      , m_interval_success(m_conf.get<script_runner::interval>(name(), "interval", m_tail ? 0s : 5s))
+      , m_interval_fail(m_conf.get<script_runner::interval>(name(), "interval-fail", m_interval_success))
+      , m_interval_if(m_conf.get<script_runner::interval>(name(), "interval-if", m_interval_success))
       , m_runner([this]() { broadcast(); }, m_conf.get(name(), "exec", ""s), m_conf.get(name(), "exec-if", ""s), m_tail,
-            m_interval, m_conf.get_with_prefix(name(), "env-")) {
+            m_interval_success, m_interval_fail, m_conf.get_with_prefix(name(), "env-")) {
     // Load configured click handlers
     m_actions[mousebtn::LEFT] = m_conf.get(name(), "click-left", ""s);
     m_actions[mousebtn::MIDDLE] = m_conf.get(name(), "click-middle", ""s);
@@ -47,7 +49,7 @@ namespace modules {
             sleep_time = m_runner.process();
           } else {
             m_runner.clear_output();
-            sleep_time = std::max(m_interval, script_runner::interval(1s));
+            sleep_time = std::max(m_interval_if, script_runner::interval(1s));
           }
 
           if (m_runner.is_stopping()) {